Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long int y;
 	float f;
 
-printf("Size of a char: %lu.\n", sizeof(c));
-printf("Size of a int: %lu.\n", sizeof(i));
-printf("Size of a long int: %lu.\n", sizeof(e));
-printf("Size of a long long int: %lu.\n", sizeof(y));
-printf("Size of a float: %lu.\n", sizeof(f));
+printf("Size of a char: %lu.\n", (unsigned long)sizeof(c));
+printf("Size of a int: %lu.\n", (unsigned long)sizeof(i));
+printf("Size of a long int: %lu.\n", (unsigned long)sizeof(e));
+printf("Size of a long long int: %lu.\n", (unsigned long)sizeof(y));
+printf("Size of a float: %lu.\n", (unsigned long)sizeof(f));
 }
 
 return (0);
